feat(ui): add table pagination
PE-847
